+++ src/game/Maps/PathFinder.cpp
@@ -345,7 +345,7 @@ void PathInfo::BuildPolyPath(const Vector3 &startPos, const Vector3 &endPos)
         // free and invalidate old path data
         clear();
 
-        unsigned int const threadId = ACE_Based::Thread::currentId();
+        unsigned int const threadId = (uintptr_t) ACE_Based::Thread::currentId();
 
         //if (threadId != m_navMeshQuery->m_owningThread)
             //sLog.outError("CRASH: We are using a dtNavMeshQuery from thread %u which belongs to thread %u!", threadId, m_navMeshQuery->m_owningThread);
